backend/geolocation/requirements.txt
# ProprScout Geolocation Service Requirements

# Core
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0

# PyTorch (CPU for compatibility, upgrade to CUDA for production)
torch>=2.0.0
torchvision>=0.15.0

# GeoCLIP and embeddings
open-clip-torch>=2.20.0
transformers>=4.35.0
sentence-transformers>=2.2.0

# Vector search
faiss-cpu>=1.7.4

# Image processing
Pillow>=10.0.0
opencv-python>=4.8.0

# GIS
geopandas>=0.14.0
shapely>=2.0.0
pyproj>=3.6.0

# Data handling
numpy>=1.24.0
pandas>=2.0.0
requests>=2.31.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
